app/requirements.txt
fastapi
uvicorn
websockets
python-multipart
numpy
faster-whisper
llama-cpp-python
edge-tts
aiofiles